feat: Add email composition function
